+++ Journal/AppDelegate.swift
@@ -88,7 +88,7 @@ extension AppDelegate: CLLocationManagerDelegate {
   }
   
   func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-    guard let location = locations.first else {
+    guard let location = locations.last else {
       return
     }